Time: 90 ms (30.14%) | Memory: 158.1 MB (73.21%) - LeetSync

+++ 3974-xor-after-range-multiplication-queries-i/xor-after-range-multiplication-queries-i.cpp
@@ -14,7 +14,7 @@ public:
             while(idx<=r){
                 long  a=nums[idx]%mod;
                 long b=v%mod;
-                long long c=(a*b)%mod;
+                long  c=(a*b)%mod;
                 nums[idx]=c;
                 idx+=k;
             }
